# backend/app/api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: DataDabble API
  description: API for creating custom databases with flexible schemas, managing teams, billing, and visualizations.
  version: "1.0.0"
  contact:
    name: DataDabble Support
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications for current user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: unread_only
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
      responses:
        "200":
          description: Paginated notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  unread_count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread notification count
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread_count:
                    type: integer
                    example: 3
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{notification_id}/read:
    parameters:
      - $ref: "#/components/parameters/NotificationId"
    put:
      tags: [Notifications]
      summary: Mark a notification as read
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification:
                    $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all notifications as read
      responses:
        "200":
          description: All marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: All notifications marked as read
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{notification_id}:
    parameters:
      - $ref: "#/components/parameters/NotificationId"
    delete:
      tags: [Notifications]
      summary: Delete a notification
      responses:
        "200":
          description: Notification deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Notification deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/preferences:
    get:
      tags: [Notifications]
      summary: Get notification preferences
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  preferences:
                    $ref: "#/components/schemas/NotificationPreference"
        "401":
          $ref: "#/components/responses/Unauthorized"

    put:
      tags: [Notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email_enabled:
                  type: boolean
                weekly_digest:
                  type: boolean
                team_invites:
                  $ref: "#/components/schemas/NotificationChannel"
                database_changes:
                  $ref: "#/components/schemas/NotificationChannel"
                entry_modifications:
                  $ref: "#/components/schemas/NotificationChannel"
                field_changes:
                  $ref: "#/components/schemas/NotificationChannel"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  preferences:
                    $ref: "#/components/schemas/NotificationPreference"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/unsubscribe/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Notifications]
      summary: Unsubscribe from email notifications (no auth required)
      security: []
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: You have been unsubscribed from email notifications.
        "404":
          description: Invalid unsubscribe link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ──────────────────────────────────────────────
  # Audit Logs
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/register

    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /api/v1/oauth2/authorize
          tokenUrl: /api/v1/oauth2/token
          scopes:
            "read:user": Read user profile
            "read:databases": Read databases
            "write:databases": Create, update, and delete databases
            "read:fields": Read fields
            "write:fields": Create, update, and delete fields
            "read:entries": Read entries
            "write:entries": Create, update, and delete entries
            "read:visualizations": Read visualizations
            "write:visualizations": Create, update, and delete visualizations
  parameters:
...
    NotificationId:
      name: notification_id
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId of the notification
      example: 665f1a2b3c4d5e6f7a8b9c11
...
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        notification_type:
          type: string
          example: database_created
        title:
          type: string
        message:
          type: string
        link:
          type: string
          nullable: true
        read:
          type: boolean
        read_at:
          type: string
          format: date-time
          nullable: true
        actor_email:
          type: string
          nullable: true
        resource_type:
          type: string
          nullable: true
        resource_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    NotificationChannel:
      type: object
      properties:
        in_app:
          type: boolean
          default: true
        email:
          type: boolean
          default: false
    NotificationPreference:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        email_enabled:
          type: boolean
        weekly_digest:
          type: boolean
        team_invites:
          $ref: "#/components/schemas/NotificationChannel"
        database_changes:
          $ref: "#/components/schemas/NotificationChannel"
        entry_modifications:
          $ref: "#/components/schemas/NotificationChannel"
        field_changes:
          $ref: "#/components/schemas/NotificationChannel"
        unsubscribe_token:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        pages:
          type: integer
          example: 3
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        messages:
          type: object
          description: Field-level validation errors
          additionalProperties: true
...
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Missing or invalid token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Resource not found